requirements.txt
pytest
playwright
pytest-playwright
appium-python-client
requests
allure-pytest
atomacos
PyYAML
pycryptodome
colorlog
pytest-xdist
python-socketio[client]
Pillow
opencv-python >= 4.10.0
numpy < 2.0.0
pyotp
pyzbar
jira
pre-commit
autoflake
isort
black == 25.9.0
flake8 == 7.3.0
